+++ src/main/java/org/threadly/concurrent/statistic/PriorityStatisticManager.java
@@ -1,4 +1,4 @@
-package org.threadly.concurrent.wrapper.statistic;
+package org.threadly.concurrent.statistic;
 
 import java.util.ArrayList;
 import java.util.Deque;
