+++ src/main/java/org/threadly/concurrent/statistic/SingleThreadSchedulerStatisticTracker.java
@@ -1,4 +1,4 @@
-package org.threadly.concurrent.wrapper.statistic;
+package org.threadly.concurrent.statistic;
 
 import java.util.ArrayList;
 import java.util.Iterator;
@@ -10,7 +10,7 @@ import java.util.concurrent.ThreadFactory;
 import org.threadly.concurrent.ConfigurableThreadFactory;
 import org.threadly.concurrent.SingleThreadScheduler;
 import org.threadly.concurrent.TaskPriority;
-import org.threadly.concurrent.wrapper.statistic.PriorityStatisticManager.TaskStatWrapper;
+import org.threadly.concurrent.statistic.PriorityStatisticManager.TaskStatWrapper;
 import org.threadly.util.Pair;
 
 /**
