+++ src/main/java/org/threadly/concurrent/statistic/package-info.java
@@ -1,0 +1,5 @@
+
+/**
+ * Package of tools and scheduler wrappers for getting statistics from concurrent systems.
+ */
+package org.threadly.concurrent.statistic;
