+++ src/main/java/org/threadly/concurrent/statistics/StatisticExecutor.java
@@ -4,12 +4,12 @@ package org.threadly.concurrent.statistics;
  * <p>Interface for some basic statistic elements provided by any statistic executor/scheduler or 
  * any executor/scheduler wrappers.</p>
  * 
- * @deprecated Moved to {@link org.threadly.concurrent.wrapper.statistic.StatisticExecutor}
+ * @deprecated Moved to {@link org.threadly.concurrent.statistic.StatisticExecutor}
  * 
  * @author jent
  * @since 4.5.0
  */
 @Deprecated
-public interface StatisticExecutor extends org.threadly.concurrent.wrapper.statistic.StatisticExecutor {
+public interface StatisticExecutor extends org.threadly.concurrent.statistic.StatisticExecutor {
   // nothing added here, this has entirely moved
 }
