+++ src/main/java/org/threadly/concurrent/wrapper/statistic/package-info.java
@@ -1,5 +1,5 @@
 
 /**
- * Package of tools and scheduler wrappers for getting statistics from concurrent systems.
+ * <p>Package of scheduler wrappers for getting statistics from concurrent systems.</p>
  */
 package org.threadly.concurrent.wrapper.statistic;
