+++ src/test/java/org/threadly/concurrent/statistic/NoThreadSchedulerStatisticTrackerTest.java
@@ -1,4 +1,4 @@
-package org.threadly.concurrent.wrapper.statistic;
+package org.threadly.concurrent.statistic;
 
 import static org.junit.Assert.*;
 import static org.threadly.TestConstants.*;
